+++ webviz_subsurface/_providers/ensemble_surface_provider/_surface_discovery.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Dict, List, Optional
 
-# The fmu.ensemble dependency ecl is only available for Linux,
+# The fmu.ensemble dependency resdata is only available for Linux,
 # hence, ignore any import exception here to make
 # it still possible to use the PvtPlugin on
 # machines with other OSes.